diff() is now part of the core API (#80)
The original code received a container full of addresses from
diff(). This was then transformed to a second container full of
value_ptr instances. This change eliminates that first container so
that we directly produce the value_ptr instances.

The diff library is now renamed dump_diff. Its only client is the
pstore-diff tool and the two are separated purely to enable unit
testing of the dump_diff code. The diff() function moves into the core
library since it's now a key component of the export functionality.

Avoid the intermediate container for diff() in emit_names().